Allow to doubletap or longpress power to toggle flashlight [1/3]
Thanks to @beanstown106 for the initial longpress action calls in
PhoneWindowManager (improved by lineage guys)

If the user enables double tap for camera action, torch action will
automatically switch to long press mode.
If the user enables double tap for torch, double tap for camera
will be disabled accordingly.

Remove the settings pref is the device doesn't have flashlight

A 170ms delay for double tap action is not annoying on single tap
to wake up the device.

Vibrate when toggling torch to let the user know of unwanted toggling
in the pocket (even if vibration is disabled by user)

Allow torch action also on ambient display

Change-Id: I7cc5849c274e9340c2f0f5fdd7e8e602f8468044