Ensure re-post with setOnlyAlertOnce(true) silences sound
Where a notification channel has an associated looping sound, it should
be possible to re-post the notification with setOnlyAlertOnce(true) to
cause the notification channel's sound to silence.

When re-posting the notification, this CL fixes an issue where
NotificationManagerService wouldn't call clearSoundLocked to stop playing
the sound, resulting in the notification continuing to playing the sound
associated with the notification channel.

Test: Manual testing, updated unit test
Bug: 110348674
Change-Id: I3ec26f0804bcf59f7356a2329b73ba8ed8f7ea51
Signed-off-by: mydongistiny <jaysonedson@gmail.com>