Don't kill process with an activity in a different visible task
The existing code only judges whether TaskRecord is in the RecentTasks
list. However it ignores other attributes of TaskRecord like isAvailable
and the flag FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS. Which means the TaskRecord
will not been displayed as a card in the RecentsActivity.

This can reduce the power of swipe up in the RecentsActivity sometimes if
the APP launched is a single task with FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS
flag. It means the APP only has one visable card in the RecentsActivity, but
when the user tries to swipe up on this card AMS will not kill the process
because the app has an extra invisable task in the RecentTasks.

https://issuetracker.google.com/issues/111857673

Change-Id: I3727ce835519a25c3ca8d80c42f285048b2e43d5
Signed-off-by: RickAi <yongbiaoai@gmail.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>